Refactor DBConnection into partial classes
